# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "easyOrder 3.0",
    "description": "Sistema de Gestão de Pedidos - Tech Challenge (3) da Pos Tech SOAT ",
    "version": "1.0.0"
  },
  "host": "",
  "paths": {
    "/cliente/cadastrar": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Cadastrar Cliente",
        "description": "Realiza o Cadastro de um novo Cliente com dados fornecidos no corpo da requisição.<br>  - Retorna o Id de Cliente como chave para continuidade no Pedido.<br>  - Não é permitido o cadastro de mais de um cliente com o mesmo CPF.<br><br>  [ Endpoint para integração ao sistema de autoatendimento ]  ",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "string",
              "required": [
                "cpf",
                "nome",
                "email"
              ],
              "properties": {
                "cpf": {
                  "type": "string",
                  "minLength": 11,
                  "maxLength": 11,
                  "example": "12345678909"
                },
                "nome": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "example": "Nome de Teste Sobrenome"
                },
                "email": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "example": "teste.email@email.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "resultado_cadastro": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Cliente cadastrado com sucesso"
                    },
                    "cliente": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
                        },
                        "nome": {
                          "type": "string",
                          "example": "João da Silva"
                        },
                        "cpf": {
                          "type": "string",
                          "example": "123.456.789-01"
                        },
                        "email": {
                          "type": "string",
                          "example": "teste@teste.com"
                        }
                      }
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "resultado_cadastro": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Cliente cadastrado com sucesso"
                    },
                    "cliente": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
                        },
                        "nome": {
                          "type": "string",
                          "example": "João da Silva"
                        },
                        "cpf": {
                          "type": "string",
                          "example": "123.456.789-01"
                        },
                        "email": {
                          "type": "string",
                          "example": "teste@teste.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Erro inesperado: Não foi possível cadastrar o cliente"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Erro inesperado: Não foi possível cadastrar o cliente"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cpf": {
                    "example": "any"
                  },
                  "nome": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cliente/atualizar": {
      "put": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualizar Cliente",
        "description": "Atualiza o Cadastro de um Cliente existente, através dos dados fornecidos no corpo da requisição.<br>  - Utiliza CPF como chave de identificação, retorna o detalhe do cadastro e Id do Cliente.<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "string",
              "required": [
                "cpf",
                "nome",
                "email"
              ],
              "properties": {
                "cpf": {
                  "type": "string",
                  "minLength": 11,
                  "maxLength": 11,
                  "example": "12345678909"
                },
                "nome": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "example": "Nome de Testes Sobrenome Atualizado"
                },
                "email": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "example": "teste-email-novo@email.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "resultado_atualizacao": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Cliente atualizado com sucesso"
                    },
                    "cliente": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
                        },
                        "nome": {
                          "type": "string",
                          "example": "João da Silva"
                        },
                        "cpf": {
                          "type": "string",
                          "example": "123.456.789-01"
                        },
                        "email": {
                          "type": "string",
                          "example": "teste@teste.com"
                        }
                      }
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "resultado_atualizacao": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Cliente atualizado com sucesso"
                    },
                    "cliente": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "29a81eeb-d16d-4d6c-a86c-e13597667307"
                        },
                        "nome": {
                          "type": "string",
                          "example": "João da Silva"
                        },
                        "cpf": {
                          "type": "string",
                          "example": "123.456.789-01"
                        },
                        "email": {
                          "type": "string",
                          "example": "teste@teste.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Erro inesperado: Não foi possível atualizar o cliente"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Erro inesperado: Não foi possível atualizar o cliente"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cpf": {
                    "example": "any"
                  },
                  "nome": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cliente/listar": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Listar Clientes",
        "description": "Lista todos os clientes cadastrados.<br><br>  [ Endpoint para integração ao sistema administrativo ]",
        "responses": {
          "200": {
            "description": "Clientes listados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Clientes listados com sucesso"
                    },
                    "clientes": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "1"
                          },
                          "cpf": {
                            "type": "string",
                            "example": "000.000.000-00"
                          },
                          "nome": {
                            "type": "string",
                            "example": "Fulano de Tal"
                          },
                          "email": {
                            "type": "string",
                            "example": "teste@teste.com"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Clientes listados com sucesso"
                    },
                    "clientes": {
                      "type": "array",
                      "items": {
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "1"
                          },
                          "cpf": {
                            "type": "string",
                            "example": "000.000.000-00"
                          },
                          "nome": {
                            "type": "string",
                            "example": "Fulano de Tal"
                          },
                          "email": {
                            "type": "string",
                            "example": "teste@teste.com"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Clientes não encontrados."
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Ocorreu um erro inesperado: Clientes não encontrados."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/cliente/buscar/{cpf}": {
      "get": {
        "tags": [
          "Descontinuadas - Mantidas para Testes Locais"
        ],
        "summary": "[Deprecated] Buscar Cliente",
        "description": "Busca Cliente por CPF.<br>  - Retorna detalhe do cadastro, incluindo o Id.<br><br>  ! Atenção: Este endpoint foi descontinuado para uso de autenticação via Lambda, no endpoint /cliente/auth/{cpf}<br><br>  [ Endpoint para integração ao sistema de autoatendimento ]",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "CPF do Cliente sem pontuação",
            "example": "12345678910"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        },
        "deprecated": true
      }
    },
    "/cliente/auth/{cpf}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "[Novo] Autenticar Cliente",
        "description": "Autenticar Cliente via CPF.<br>  - Retorna o Id do cliente, caso encontrado, como chave para continuidade no Pedido.<br><br>  [ Endpoint para integração ao sistema de autoatendimento ]",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "CPF do Cliente sem pontuação",
            "example": "12345678910"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}